:sparkles: Added new tool 'Tick Price' | 'Hook Deployer' is coming soon.
