+++ app/tools/page.jsx
@@ -56,17 +56,16 @@ export default async function Page() {
       <HeroBanner
         title="Tools"
         subtitle="Tools to help you understand Uniswap v4"
-      ></HeroBanner>
+      />
 
       <Container classNames="pb-8 lg:pb-12">
-        {toolPosts.map((toolPost) => (
-          <li
-            key={toolPost.id}
-            className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4"
-          >
-            <CollectionCard componentData={toolPost} />
-          </li>
-        ))}
+        <ul className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
+          {toolPosts.map((toolPost) => (
+            <li key={toolPost.id} className="space-y-4">
+              <CollectionCard componentData={toolPost} />
+            </li>
+          ))}
+        </ul>
       </Container>
     </>
   )
