+++ src/data/tools/hook-deployer.mdx
@@ -1,0 +1,9 @@
+---
+id: hook-deployer
+title: Hook Deployer
+description: Uniswap V4 Hook create2 deployer, for use with Uniswap V3.
+docs: https://github.com/drgorillamd/uniV4-hook-deployer
+tag: soon
+emoji: 🛜
+category: tools
+---
