+++ src/data/tools/square-root-price.mdx
@@ -1,6 +1,6 @@
 ---
-id: uniswap-v4-prices
-title: Uniswap V4 Prices
+id: square-root-price
+title: Square Root Price
 description: Calculate Uniswap V4 ratios (prices) from token pairs.
 docs: https://github.com/thanpolas/univ3prices/blob/main/README.md
 tag: new
