+++ src/data/tools/tick-price.mdx
@@ -1,0 +1,9 @@
+---
+id: tick-price
+title: Tick Price
+description: Calculates Liquidity Pool (LP) ratios (prices) for Token Pairs using the current tick value.
+docs: https://github.com/thanpolas/univ3prices/blob/main/README.md#univ3pricestickpricetokendecimals-tick
+tag: new
+emoji: 🧮
+category: tools
+---
